+++ packages/integrations/mocked/src/modules/cms/mappers/blocks/cms.order-list.mapper.ts
@@ -91,6 +91,11 @@ const MOCK_ORDER_LIST_BLOCK_EN: CMS.Model.OrderListBlock.OrderListBlock = {
                     label: 'To',
                 },
             },
+            {
+                __typename: 'FilterViewModeToggle',
+                id: 'viewMode',
+                value: 'list',
+            },
         ],
     },
     noResults: {
@@ -105,6 +110,11 @@ const MOCK_ORDER_LIST_BLOCK_EN: CMS.Model.OrderListBlock.OrderListBlock = {
     },
     reorderLabel: 'Reorder',
     detailsUrl: '/orders/{id}',
+    cardHeaderSlots: {
+        left: 'id',
+        right: 'status',
+        bottom: 'createdAt',
+    },
 };
 
 const MOCK_ORDER_LIST_BLOCK_DE: CMS.Model.OrderListBlock.OrderListBlock = {
@@ -198,6 +208,11 @@ const MOCK_ORDER_LIST_BLOCK_DE: CMS.Model.OrderListBlock.OrderListBlock = {
                     label: 'Bis',
                 },
             },
+            {
+                __typename: 'FilterViewModeToggle',
+                id: 'viewMode',
+                value: 'list',
+            },
         ],
     },
     noResults: {
@@ -212,6 +227,11 @@ const MOCK_ORDER_LIST_BLOCK_DE: CMS.Model.OrderListBlock.OrderListBlock = {
     },
     reorderLabel: 'Erneut bestellen',
     detailsUrl: '/bestellungen/{id}',
+    cardHeaderSlots: {
+        left: 'id',
+        right: 'status',
+        bottom: 'createdAt',
+    },
 };
 
 const MOCK_ORDER_LIST_BLOCK_PL: CMS.Model.OrderListBlock.OrderListBlock = {
@@ -305,6 +325,11 @@ const MOCK_ORDER_LIST_BLOCK_PL: CMS.Model.OrderListBlock.OrderListBlock = {
                     label: 'Do',
                 },
             },
+            {
+                __typename: 'FilterViewModeToggle',
+                id: 'viewMode',
+                value: 'list',
+            },
         ],
     },
     noResults: {
@@ -319,6 +344,11 @@ const MOCK_ORDER_LIST_BLOCK_PL: CMS.Model.OrderListBlock.OrderListBlock = {
     },
     reorderLabel: 'Zamów ponownie',
     detailsUrl: '/zamowienia/{id}',
+    cardHeaderSlots: {
+        left: 'id',
+        right: 'status',
+        bottom: 'createdAt',
+    },
 };
 
 export const mapOrderListBlock = (locale: string): CMS.Model.OrderListBlock.OrderListBlock => {