+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @o2s/telemetry
 
+## 1.1.1
+
+### Patch Changes
+
+- 0354126: reworked turbo dev task to optimize CPU/RAM usage
+
 ## 1.1.0
 
 ### Minor Changes